Add debug log

+++ src/nameserver/block_mapping.h
@@ -44,7 +44,7 @@ public:
     void DealWithDeadBlocks(int64_t cs_id, const std::set<int64_t>& blocks);
     bool SetBlockVersion(int64_t block_id, int64_t version);
     int32_t PickRecoverBlocks(int32_t cs_id, int32_t block_num, std::map<int64_t, int32_t>* recover_blocks);
-    void ProcessRecoveredBlock(int64_t cs_id, int64_t block_id, bool recover_success);
+    void ProcessRecoveredBlock(int32_t cs_id, int64_t block_id, bool recover_success);
     void GetStat(int64_t* recover_num, int64_t* pending_num);
 
 private:
